added dtype to remove annoying warning message

+++ envs/json_walker.py
@@ -135,8 +135,8 @@ class JSONWalker(gym.Env):
         high = np.array( [np.inf]*(5*len(self.body_defs)+2*len(self.joint_defs)+10) )
 
         self.action_space = spaces.Box(
-            np.array([-1.0]*num_enabled_joints), np.array([+1.0]*num_enabled_joints))
-        self.observation_space = spaces.Box(-high, high)
+            np.array([-1.0]*num_enabled_joints), np.array([+1.0]*num_enabled_joints), dtype=np.float32)
+        self.observation_space = spaces.Box(-high, high, dtype=np.float32)
 
     def seed(self, seed=None):
         self.np_random, seed = seeding.np_random(seed)
